requirements.txt
annotated-types==0.7.0
anyio==4.5.2
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
exceptiongroup==1.2.2
greenlet==3.1.1
h11==0.14.0
idna==3.10
mysql-connector-python==9.0.0
pycparser==2.22
pydantic==2.10.1
pydantic-core==2.27.1
PyMySQL==1.1.1
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.41.3
typing-extensions==4.12.2
uvicorn==0.32.1
